Update serial examples to also allow input
