+++ examples/daplink-serial/usb.js
@@ -24,6 +24,7 @@ const usb = require("usb");
 const common = require("./common");
 const DAPjs = require("../../");
 
+// Read USB device descriptor
 function getStringDescriptor(device, index) {
     return new Promise((resolve, reject) => {
         try {
@@ -49,20 +50,18 @@ function selectDevice(vendorID) {
             return reject("No devices found");
         }
 
-        process.stdin.setRawMode(true);
-        process.stdin.setEncoding("utf8");
-        process.stdin.on("readable", () => {
-            let input = process.stdin.read();
+        function inputHandler(input) {
             if (input === "\u0003") {
                 process.exit();
             }
 
             let index = parseInt(input);
             if (index && index <= devices.length) {
-                process.stdin.setRawMode(false);
+                common.inputEmitter.removeListener("input", inputHandler);
                 resolve(devices[index - 1]);
             }
-        });
+        }
+        common.inputEmitter.addListener("input", inputHandler);
 
         console.log("Select a device to listen to serial output:");
         devices.forEach((device, index) => {
