+++ examples/read-registers/webusb.js
@@ -26,19 +26,8 @@ const DAPjs = require("../../");
 
 // Allow user to select a device
 function handleDevicesFound(devices, selectFn) {
-    process.stdin.setRawMode(true);
-    process.stdin.setEncoding("utf8");
-    process.stdin.on("readable", () => {
-        let input = process.stdin.read();
-        if (input === "\u0003") {
-            process.exit();
-        }
-
-        let index = parseInt(input);
-        if (index && index <= devices.length) {
-            process.stdin.setRawMode(false);
-            selectFn(devices[index - 1]);
-        }
+    common.inputEmitter.addListener("input", index => {
+        if (index <= devices.length) selectFn(devices[index - 1]);
     });
 
     console.log("Select a device to read registers:");
